+++ typescript/sdk/src/ism/HyperlaneIsmFactory.ts
@@ -24,9 +24,13 @@ import {
 } from '@hyperlane-xyz/core';
 import {
   Address,
+  Domain,
   eqAddress,
   formatMessage,
   normalizeAddress,
+  objFilter,
+  objMap,
+  warn,
 } from '@hyperlane-xyz/utils';
 
 import { HyperlaneApp } from '../app/HyperlaneApp';
@@ -214,6 +218,22 @@ export class HyperlaneIsmFactory extends HyperlaneApp<ProxyFactoryFactories> {
     const overrides = this.multiProvider.getTransactionOverrides(destination);
     const routingIsmFactory = this.getContracts(destination).routingIsmFactory;
     let routingIsm: DomainRoutingIsm | DefaultFallbackRoutingIsm;
+    // filtering out domains which are not part of the multiprovider
+    config.domains = objFilter(
+      config.domains,
+      (domain, config): config is IsmConfig => {
+        const domainId = this.multiProvider.tryGetDomainId(domain);
+        if (domainId === null) {
+          warn(
+            `Domain ${domain} doesn't have chain metadata provided, skipping ...`,
+          );
+        }
+        return domainId !== null;
+      },
+    );
+    const safeConfigDomains = Object.keys(config.domains).map((domain) =>
+      this.multiProvider.getDomainId(domain),
+    );
     const delta: RoutingIsmDelta = existingIsmAddress
       ? await routingModuleDelta(
           destination,
@@ -225,7 +245,7 @@ export class HyperlaneIsmFactory extends HyperlaneApp<ProxyFactoryFactories> {
         )
       : {
           domainsToUnenroll: [],
-          domainsToEnroll: Object.keys(config.domains),
+          domainsToEnroll: safeConfigDomains,
         };
 
     const signer = this.multiProvider.getSigner(destination);
@@ -240,13 +260,14 @@ export class HyperlaneIsmFactory extends HyperlaneApp<ProxyFactoryFactories> {
     }
     // reconfiguring existing routing ISM
     if (existingIsmAddress && isOwner && !delta.mailbox) {
-      const isms: ChainMap<Address> = {};
+      const isms: Record<Domain, Address> = {};
       routingIsm = DomainRoutingIsm__factory.connect(
         existingIsmAddress,
         this.multiProvider.getSigner(destination),
       );
       // deploying all the ISMs which have to be updated
-      for (const origin of delta.domainsToEnroll) {
+      for (const originDomain of delta.domainsToEnroll) {
+        const origin = this.multiProvider.getChainName(originDomain); // already filtered to only include domains in the multiprovider
         logger(
           `Reconfiguring preexisting routing ISM at for origin ${origin}...`,
         );
@@ -256,23 +277,20 @@ export class HyperlaneIsmFactory extends HyperlaneApp<ProxyFactoryFactories> {
           origin,
           mailbox,
         });
-        isms[origin] = ism.address;
+        isms[originDomain] = ism.address;
         const tx = await routingIsm.set(
-          this.multiProvider.getDomainId(origin),
-          isms[origin],
+          originDomain,
+          isms[originDomain],
           overrides,
         );
         await this.multiProvider.handleTx(destination, tx);
       }
       // unenrolling domains if needed
-      for (const origin of delta.domainsToUnenroll) {
+      for (const originDomain of delta.domainsToUnenroll) {
         logger(
-          `Unenrolling origin ${origin} from preexisting routing ISM at ${existingIsmAddress}...`,
+          `Unenrolling originDomain ${originDomain} from preexisting routing ISM at ${existingIsmAddress}...`,
         );
-        const tx = await routingIsm.remove(
-          this.multiProvider.getDomainId(origin),
-          overrides,
-        );
+        const tx = await routingIsm.remove(originDomain, overrides);
         await this.multiProvider.handleTx(destination, tx);
       }
       // transfer ownership if needed
@@ -293,9 +311,6 @@ export class HyperlaneIsmFactory extends HyperlaneApp<ProxyFactoryFactories> {
         isms[origin] = ism.address;
       }
       const submoduleAddresses = Object.values(isms);
-      const domains = Object.keys(isms).map((chain) =>
-        this.multiProvider.getDomainId(chain),
-      );
       let receipt: ethers.providers.TransactionReceipt;
       if (config.type === IsmType.FALLBACK_ROUTING) {
         // deploying new fallback routing ISM
@@ -315,7 +330,7 @@ export class HyperlaneIsmFactory extends HyperlaneApp<ProxyFactoryFactories> {
           destination,
           routingIsm['initialize(address,uint32[],address[])'](
             config.owner,
-            domains,
+            safeConfigDomains,
             submoduleAddresses,
             overrides,
           ),
@@ -324,7 +339,7 @@ export class HyperlaneIsmFactory extends HyperlaneApp<ProxyFactoryFactories> {
         // deploying new domain routing ISM
         const tx = await routingIsmFactory.deploy(
           config.owner,
-          domains,
+          safeConfigDomains,
           submoduleAddresses,
           overrides,
         );
@@ -441,12 +456,17 @@ export async function moduleCanCertainlyVerify(
   origin: ChainName,
   destination: ChainName,
 ): Promise<boolean> {
+  const originDomainId = multiProvider.tryGetDomainId(origin);
+  const destinationDomainId = multiProvider.tryGetDomainId(destination);
+  if (!originDomainId || !destinationDomainId) {
+    return false;
+  }
   const message = formatMessage(
     0,
     0,
-    multiProvider.getDomainId(origin),
+    originDomainId,
     ethers.constants.AddressZero,
-    multiProvider.getDomainId(destination),
+    destinationDomainId,
     ethers.constants.AddressZero,
     '0x',
   );
@@ -713,7 +733,11 @@ export async function routingModuleDelta(
   const routingIsm = DomainRoutingIsm__factory.connect(moduleAddress, provider);
   const owner = await routingIsm.owner();
   const deployedDomains = (await routingIsm.domains()).map((domain) =>
-    multiProvider.getChainName(domain.toNumber()),
+    domain.toNumber(),
+  );
+  // config.domains is already filtered to only include domains in the multiprovider
+  const safeConfigDomains = objMap(config.domains, (domain) =>
+    multiProvider.getDomainId(domain),
   );
 
   const delta: RoutingIsmDelta = {
@@ -731,16 +755,15 @@ export async function routingModuleDelta(
   }
   // check for exclusion of domains in the config
   delta.domainsToUnenroll = deployedDomains.filter(
-    (domain) => !Object.keys(config.domains).includes(domain),
+    (domain) => !Object.values(safeConfigDomains).includes(domain),
   );
   // check for inclusion of domains in the config
   for (const [origin, subConfig] of Object.entries(config.domains)) {
-    if (!deployedDomains.includes(origin)) {
-      delta.domainsToEnroll.push(origin);
+    const originDomain = safeConfigDomains[origin];
+    if (!deployedDomains.includes(originDomain)) {
+      delta.domainsToEnroll.push(originDomain);
     } else {
-      const subModule = await routingIsm.module(
-        multiProvider.getDomainId(origin),
-      );
+      const subModule = await routingIsm.module(originDomain);
       // Recursively check that the submodule for each configured
       // domain matches the submodule config.
       const subModuleMatches = await moduleMatchesConfig(
@@ -749,9 +772,9 @@ export async function routingModuleDelta(
         subConfig,
         multiProvider,
         contracts,
-        origin,
+        mailbox,
       );
-      if (!subModuleMatches) delta.domainsToEnroll.push(origin);
+      if (!subModuleMatches) delta.domainsToEnroll.push(originDomain);
     }
   }
   return delta;